+++ emergencyRoom/src/main/java/com/brettistagge/IssueProcedure.java
@@ -14,7 +14,7 @@ import javax.validation.constraints.NotNull;
 public class IssueProcedure {
     private @Id @GeneratedValue long issueProcedureId;
     private @NotNull @Column(unique = true) String name;
-    private @NotNull long cost;
+    private @NotNull Integer cost;
 
     @ManyToOne
     @JoinColumn(name = "issue")
